none

+++ SMG_NothingIsAll/Demo/BirdLivePage/View/BirdView.h
@@ -53,7 +53,7 @@
  *  MARK:--------------------摸翅膀--------------------
  *  @param direction 从左顺时针,8个方向,分别为0-7;
  */
--(void) touchWing:(int)direction;
+-(void) touchWing:(long)direction;
 
 /**
  *  MARK:--------------------痛--------------------
