+++ SMG_NothingIsAll/Demo/BirdLivePage/View/BirdView.m
@@ -61,7 +61,7 @@
     CGFloat angle = value_F1_1 * M_PI;
     
     //4. 用sin计算对边Y,cos计算邻边X;
-    NSLog(@"fly >> %@ angle:%f",[NVHeUtil getLightStr_Value:value algsType:FLY_RDS dataSource:@""],value_F1_1 * 180);
+    NSLog(@"fly >> %@ angle:%.0f",[NVHeUtil getLightStr_Value:value algsType:FLY_RDS dataSource:@""],value_F1_1 * 180);
     [UIView animateWithDuration:0.1f animations:^{
         [self setX:self.x + (cos(angle) * 30.0f)];
         [self setY:self.y + (sin(angle) * 30.0f)];
@@ -100,7 +100,7 @@
  *  MARK:--------------------摸翅膀--------------------
  *  @param direction 从左顺时针,8个方向,分别为0-7;
  */
--(void) touchWing:(int)direction{
+-(void) touchWing:(long)direction {
     //1. 飞前视觉
     //[self see:[self.delegate birdView_GetPageView]];
